+++ LootMethods/WhitelistLoot.cs
@@ -6,10 +6,10 @@ namespace LootManager
         {
             if (item == null || string.IsNullOrWhiteSpace(item.ItemName))
                 return false;
-            
+
             if (item.NoTradeNoDestroy)
                 return true;
-            
+
             if (Plugin.LootEquipment.Value && item.RequiredSlot != Item.SlotType.General)
             {
                 switch (Plugin.LootEquipmentTier.Value)
@@ -18,20 +18,24 @@ namespace LootManager
                         return true;
                     case EquipmentTierSetting.NormalOnly:
                         return quantity == 1;
+                    case EquipmentTierSetting.ImprovedOnly:
+                        return quantity > 10;
                     case EquipmentTierSetting.BlessedOnly:
                         return quantity == 2;
                     case EquipmentTierSetting.AscendedOnly:
                         return quantity == 3;
-                    case EquipmentTierSetting.BlessedAndUp:
+                    case EquipmentTierSetting.ImprovedAndUp:
                         return quantity >= 2;
+                    case EquipmentTierSetting.BlessedAndUp:
+                        return quantity >= 2 && quantity <= 3;
                 }
             }
 
             string itemName = item.ItemName;
-            
+
             if (Plugin.Whitelist.Contains(itemName))
                 return true;
-            
+
             foreach (var kvp in Plugin.FilterList)
             {
                 if (!Plugin.EnabledFilterCategories.Contains(kvp.Key))  continue;