+++ Patches/LootWindowPatch.cs
@@ -45,7 +45,7 @@ namespace LootManager
                 }
                 
 
-                if (Plugin.Auctionlist != null && Plugin.Auctionlist.Contains(name))
+                if (Plugin.AuctionLootEnabled.Value && Plugin.Auctionlist != null && Plugin.Auctionlist.Contains(name))
                 {
                     if (AuctionLoot.TryListItem(item, qty))
                     {